Fix writing amf3 reference tables in combinators
In several places, we updated a reference table RefCell,
and then returned a SerializeFn. However, the returned function
might not actually be executed (e.g. when used with 'cond'),
resulting in elements incorrectly being added to a reference table.

We now update all reference tables from within the 'SerializeFn'
functions we create. This fixes LearnToFly3