Disable broken object reference table writing for now
We currently deduplicate objects in the object reference table
with a derived PartialEq implementation, which compares everything
by value. However, this will incorrectly cause us to use the same
object reference table entry for two distinct objects or vectors
which happen to have the same values. This is observable in
Actionscript, as the same ActionScript object will be used for
both.

Until we can fix this, let's never re-use a slot in the object
reference table (we avoid using 'to_length_store'). While this
will result in a re-use of the same 'Rc<Value>' being serialized
as two distinct objects, we can detect about warn about this from
the Ruffle side. This is much less likely to cause issues while
we're working on a fix, since it's common to have multiple objects
with the value (e.g. serializing multiple empty 'save slot' objects
in LearnToFly3).

The previous LearnToFly3 test was created using the buggy
implementation. I've replaced it with data derived directly
from Flash Player.